Split ApacheOphHttpClient to separate classes
