+++ README.md
@@ -4,7 +4,7 @@ Yleiskäyttöisiä java-kirjastoja Opetushallituksen verkkokehitykseen
 
 ## Kehitys
 
-* JDK 1.7, koska osa käyttävistä projekteista on vielä 1.7
+* Käytössä JDK 1.8
 * Jokainen alimoduli julkaisee oman jar-pakettinsa
 * Alimodulit mahdollisimman yksinkertaisina: [SRP](https://en.wikipedia.org/wiki/Single_responsibility_principle)
 * Lisää uudet java-luokat omiin alimoduleihinsa, varsinkin jos liity kiinteästi olemassaolevaan pakettiin
