+++ httpclient/src/main/java/fi/vm/sade/javautils/httpclient/OphHttpClientProxy.java
@@ -1,5 +1,5 @@
 package fi.vm.sade.javautils.httpclient;
 
 public abstract class OphHttpClientProxy extends OphRequestParameterAccessors<OphHttpClientProxy> implements AutoCloseable {
-    abstract OphHttpClientProxyRequest createRequest(OphRequestParameters requestParameters);
+    public abstract OphHttpClientProxyRequest createRequest(OphRequestParameters requestParameters);
 }
